add .env file existance check to bot.py

+++ bot.py
@@ -1,5 +1,5 @@
 import asyncio
-from os import getenv
+from os import getenv, path
 
 from aiogram import Bot, Dispatcher
 from aiogram.client.default import DefaultBotProperties
@@ -10,12 +10,19 @@ from app.handlers import router
 
 
 async def main() -> None:
-    load_dotenv()
-    bot = Bot(token=getenv("API_TOKEN"), default=DefaultBotProperties(parse_mode=ParseMode.HTML))
-    dp = Dispatcher()
-    dp.include_router(router)
-    await bot.delete_webhook(drop_pending_updates=True)
-    await dp.start_polling(bot)
+    if path.isfile(".env"):
+        load_dotenv()
+        bot = Bot(token=getenv("API_TOKEN"), default=DefaultBotProperties(parse_mode=ParseMode.HTML))
+        dp = Dispatcher()
+        dp.include_router(router)
+        await bot.delete_webhook(drop_pending_updates=True)
+        await dp.start_polling(bot)
+    else:
+        print("Файл .env не найден!\n"
+              "1. Создайте файл с именем \".env\" рядом с файлом \"bot.py\"\n"
+              "2. В файл \".env\" добавьте строку:\n"
+              "API_TOKEN = \"telegram_bot_token\"\n"
+              "где telegram_bot_token - это токен для вашего телеграм-бота, полученный от @BotFather")
 
 
 if __name__ == "__main__":
